Re-enable the animation of animated images
This used to work, but has been lost at some point.

The reason is that the call to gdk_pixbuf_animation_iter_get_pixbuf()
does not follow the specifications, which state: " Copy the pixbuf to
keep it (don't just add a reference), as it may get recycled as you
advance the iterator."

So, make a copy, since we don't own a reference to the returned pixbuf
anyway.

This fixes https://sourceforge.net/p/lxde/bugs/814,
https://bugs.launchpad.net/ubuntu/+source/gpicview/+bug/1448561 and
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1010719.