Fix a bug with the automatically enlarged window on opening
If an image is larger than the default window size but still fits on the
screen, the default window size is automatically increased to display
the image without zooming. However, the calculation is based on a window
in the upper left corner.

Take into account that the user can now set the position of the window
by themselves.