Fix a bug with the automatically enlarged window on opening
If an image is larger than the default window size but still fits on the
screen, the default window size is automatically increased to display
the image without zooming. With GTK+ 2, however, scrollbars appear even
though the image is displayed without zooming.

Let GTK+ 2 "realize" the window changes before continuing.